Move last bits of dashboard.ts to separate modules
